always return OK to the user
This will always use a 200 response type. The
assumed use-case is a proxy that tees the
requests to both a typical publish endpoint and
this cluster. The user response should come
from the publish endpoint with the response
here merely acknowledging receipt. For now to
avoid spurious retries and confusion this
response is always OK.