fix auto-reconnect for configbin source (#19)
Moves the recover logic for the entity source to
the correct place. Also changes the retry source
to flatMap the source for a single connect including
the data bytes so the retry will be backpressured
until the request fails overall. Before it could
continue to make requests as soon as a response
was received.