adjust values to fit CloudWatch restrictions (#28)
To reduce the number of put calls, the sends to CloudWatch
are batched. The downside is that if a value is invalid, then
the whole batch gets rejected. In particular, we were seeing
NaN values cause the batch to fail with:

```
com.amazonaws.services.cloudwatch.model.InvalidParameterValueException:
  The value ? for parameter MetricData.member.1.Value is invalid.
  (Service: AmazonCloudWatch; Status Code: 400;
   Error Code: InvalidParameterValue; Request ID: XYZ)
```

This change adds a truncate step that coerces the value to
fit the CloudWatch restrictions. The truncation logic was
borrowed from the Servo library and has been used for many
years in that library.