update the expr groups to include app and name
The vast majority of the evaluated expressions we have
internally have an exact match on the app and the name.
With this change the bridge will group the expression
set by both in the evaluator to reduce the number of
expressions that have to be checked for any given datapoint.

On an internal test node with real traffic this improved
RPS by about 10x.