use QueryIndex for LWC bridge
Refactors the expression evaluation for the LWC bridge
to levarage the Atlas query index. Before it was only
segmenting by application. If a larger portion of the
queries due not have an exact match on application,
cluster, or asg, then we cannot extract an application
and they would get placed in the catch all bucket.
Expressions in the catch all must be checked against
all datapoints flowing through the system.

With the query index we can drill down based on any
equal clauses in the query. For the actual data set
that was causing problems there are 18.5k expressions
and 2500 were in the catch all bucket. Using the query
index that top-level fallback set reduced to 46. As
a further optimization we are now ignoring those top-level
fallbacks from the index and requiring the user to be
more restrictive on the query.