fix filtering for druid tag queries
We use the `nf.datasource` tag to specify the Druid
datasource and the `name` to match a metric for that
datasource. When searching for the list of values for
a given dimension, we need to ensure those aspects of
the Atlas query are removed before converting the
remainder of the query to a Druid filter.