convert druid values to rate per second (#45)
For now all druid data is assumed to be counters and the
values we get back are a rate per step. This changes them
to a rate per second to be consistent with Atlas sources.
We may need to revisit in the future, but we'll need to
figure out how to automatically understand the typing on
the druid side first.