initial version of aggregator service
This is an initial prototype for an aggregator service
to receive deltas from stateless clients. The api is
similar to internal aggregator that has been used for
a while to perform inline removal of the node dimension.
In this case updates are forwarded to an AtlasRegistry
so it will behave like a normal client on behalf of
the clients.

For more details see Netflix/spectator#432.

Left as future work:

- Reduce the expiration period for the Atlas registry.
  This will require changes on the Spectator side to
  improve the cleanup.
- Minimal validation of datapoints. In particular checks
  for tag length and number of tags should be performed.
- Automatic rollup based on cardinality or churn. This
  is to protect the backends and avoid explosion of metrics
  volume if improper tagging is used. In particular with
  FaaS systems there is a concern about dimensions that
  might have a request id. Ideally, rollups will apply to
  the Atlas publish path, but not the LWC path. This would
  allow users to get full dimensions on-demand via
  streaming.
- Check how resilient we are to DDoS. A concern with this
  approach is poorly written clients may send a delta for
  every counter update in a tight loop. For big clusters
  this could be problematic.