Fix notification structure
I had initially thought I would pass the `Message` portion of the
notification, but that didn't turn out to be the case. In the meantime,
that incorrect assumption was encoded into the tests.

This commit makes the structure of the test messages match what is seen
in the live service and fixes the logic that extracts information for
Atlas metrics.