avoid entity size limit for fetch expressions
The default limit is 8m and we have exceeded that in some
environments. Also, the current behavior seems to be it will
hang the stream with no clear error if this size is exceeded.
So for now we'll remove the size limit so at least the failure
mode will be more obvious.