druid: add range query keys to group by dimensions (#111)
Keys used as part of a query clause that can return a set of
values need to be included as part of the result set so we can
perform the local evaluation step. Before this change the user
would incorrectly get an empty result if `:re` or similar ops
were used with a key that was not included in the explicit set
of group by keys.