use segmentMetadata for describing data sources
Switches to using segmentMetadata queries for understanding
the set of dimensions and metrics for a given data source.
This has two benefits: 1) it seems to be more reliable as
in some error cases the GET request to /datasources doesn't
show any dimensions, and 2) it gives us type information
that can be used to make a more informed choice about how
to query the data.

The main downside is that segmentMetadata seems to be quite
a bit more expensive to compute. The intervals are limited
to the tag query interval for now which seems to keep the
overhead to a reasonable level.