Add endpoint for removing purge eligible expressions (#128)
* Add endpoint for removing purge eligible expressions

Adds an endpoint to update or remove cluster configurations
in ConfigBin for the given expressions.

Modified the cluster config schema to allow empty list
for `checksToSkip` field. This is convenient because Json
encoding creates this empty list.

Moved ConfigBinVersion to the shared library.

* Address review comments

Add more contextual comment when removing expressions.

Always update the Cluster config instead of deleting to
preserve the history.

Remove the entry from DynamoDB when the expression is
purged.