druid: filter out datapoints with null values (#129)
Druid treats null and empty string as being the same. If
dimensions returned as part of the event map from a group
by request have a null value, then filter them out to avoid
NPE during request processing in Atlas. For the Atlas storage
null values are not allowed and cannot be part of the response.