Atlas Slotting Service (#132)
Given a set of app names, the service periodically polls AWS AutoScalingGroups
(ASGs) and AWS EC2 Instances, so that it can assign slot numbers to the
instances, from 0 to N-1. As instances in the ASG are replaced, they take over
missing slot numbers on a first-come, first-served basis. To the extent that
instances in the ASG remain stable, so will the slot numbers.

Slotting information is used by Atlas in the hashing scheme which divides
data between nodes in large backend clusters. This information is one of the
pre-requisites for running large Atlas clusters.

See README.md for more details.