DynamoDB Capacity Updates
The required DynamoDB capacity can differ based on account and region,
depending on the number and size of ASGs being tracked. This change
provides the ability to specify the desired read and write capacity per
account and region, with a default value as the fallback.

The timestamp update and deactivate ASG methods do not change the data
payload in DynamoDB, so the condition expressions have been updated to
check the data that is changing. the data payloads can be up to 400KB in
size, while a Boolean and a Long are just a few bytes. this should help
save some DynamoDB write capacity. keeping the conditional checks in
place will provide stability for the items, when there are multiple
writers during a deployment.