Scaling policy lookup for expressions (#130)
Query Edda to determine whether a scaling policy is
created in Ec2 or Titus for the metric that is forwarded
to CW.

Also store list of forwarded metrics info and scaling
policies instead of just one. A list is necessary when
multiple metric lines are forwarded because of a
group-by query.