Log SES notifications as structured JSON
An internal Netflix service that will consume these notifications
requires the notification to be structured JSON (as opposed to string
encoded JSON). This commit changes the logging flow to accommodate that
requirement. In addition to logging structured JSON, the raw SQS message
is now logged to the primary log file for diagnostics.