Migrate to iep.aws2.AtlasClientFactory
No async client was available in the factory for the AWS 1.x SDK and,
since the SES Monitor usage was a one-off, I didn't add it. Now that
alpakka-sqs uses the AWS 2.x SDK and there is a client factory for that
available, I'm moving the SES Monitor to leverage it.