Purge the forwarded metrics admin record that are no longer reported (#156)
Save the report timestamp for every forwarded metric and use it
to remove the entry when the metric is no longer reported.

This is useful to keep the list size bounded for the forwarded metrics
and scaling policies in a 'ExpressionDetails' admin record.
When one expression is used to forward metrics for many asgs,
the admin record should be updated when the asgs are removed and
not forwarding any data.

Some more changes:
Use the mapped env for the account if one is configured. Jenkins
clusters use hardcoded account which requires explicit mapping
for the environment.

Added timers for calls to DynamoDB

Throttle the flow to avoid bursts of requests to Dynamodb