druid: fix `key,:has,:not` queries
Before these would be empty because druid would return the
event with a `"key": null` and it would get filtered out.
The filtering was added to avoid NPE with percentile queries.

This changes it to filter the keys with null values from the
set of tags rather than the overall datapoint. Checking with
a percentile query, this does not regress that behavior.