druid: limit dimensions in grouping
Before any range query would get included in the dimension
list so that local filtering would work correctly. It is now
limited to keys that are explicitly in the group by. The
local filtering should still avoid problems with multi-value
keys that are included in the group by. The eval step will
only check the exact key matches to avoid over filtering
based on context that is no longer part of the result.