use multiple AtlasRegistry instances for aggregator
The AtlasRegistry is generally meant for typical app usage
and tries to limit the amount of resources it takes. For
example in many cases it will limit the usage to a single
thread to bound the total compute it could possibly take.
For the aggregator, we want to use more of the machine
resources. In this change we partition the data across a
set of AtlasRegistry instances to allow for more concurrent
processing of the data.