validate datapoints coming into the aggregator (#182)
The `/update` endpoint on the aggregator will now validate
the datapoints and indicate any issues in the response. This
can be used by thin clients to inform the user of problems.

The rules are mostly the same as those used for the publish
endpoint of the backend. Valid characters are not checked
because they will get fixed on this path by the registry.