druid: do not limit search query response size (#212)
Before we would push the limit for the number of results
to the search query as an optimization to reduce the amount
of data coming back from Druid to only what was needed.
The problem is that for multi-value dimensions if any value
for a given record matches, then all values for that record
will get returned. Those will get filtered locally, but with
a limit on the Druid side, the value we are looking for may
be past that limit and cause an empty result set.