aggr: adjust initial polling delay
Adjust the initial polling delay on the aggregator
to help spread out the requests to the downstream
clusters. It was initially aligned to the start to
maximize the amount of time to send the data. With
recent changes to run larger aggregator clusters
that have less data to send per node, this can
create a thundering herd for the connections.