persistence: bucket files by start time to make range search easier (#254)
Persistence: put files with the same start/end minute into a sub folder in s3, to make file matching by range easier for downstream data processing pipeline. 
In theory it can create up to 1830(60+59+...+1) folders for an hour, but in reality I'd expect actual number to be much smaller because data points are mostly in order and range for a file is small, e.g. assume max range per file is 3 min, max num of folders <= 3*60.