include tags in validation responses (#258)
For more information see Netflix/atlas#1203. It might be
possible to reduce some of the duplication with PublishApi,
but need to ensure it doesn't cause problems and that can
be done at a later time if needed. Currently aggregator
does some of the validation as parsing to reduce the overhead.