bridge: switch to spectator QueryIndex impl
Simplify the LWC bridge by using the Spectator implementation
of the QueryIndex. It can take advantage of the sorted ids
and mutations to reduce the overhead considerably. However,
to efficiently get the data in the right form to feed into
it, we need to change the parsing so we avoid creating the
intermediate Datapoint objects that are returned by
PublishApi.