add a metric to track tmp files copied to s3 (#284)
So far the only times we’ve seen corrupt files on s3 is that
tmp files are not closed properly and get copied. This counter
should allow us to track those occurrences. If the file is closed
properly and is corrupt for some other reason, then it will not
detect those instances.