lwc bridge: add metric for payload size (#296)
Adds a metric to track the raw payload size for the set
of expressions to sync from lwcapi. Also converts the
number of expressions metric to a matching distribution
summary.

Minor cleanup: 1) inject actor system, and 2) avoid
creating a copy of the response as a byte array to parse
the data.