fix vpn metrics
The VpnId and and TunnelIpAddress dimensions are not both present on
the metrics that are published - it is one or the other. Each VPN
connection has two tunnels, a primary and a backup, and the preference
is to monitor these metrics based on VPN connection. As a result, we
will drop the TunnelIpAddress dimension.