reduce overhead for encoding payloads
Write directly to the json generator rather than using
the mapper. Since this is a hot path and the objects
are fairly simple this change helps reduce the overhead
for encoding the payloads.

Also refactors so the encoding is not part of the entity
source. This helps make it clear whether the latency is
due to the encoding or the actual network call.