add aws.memorydb tag
This maps the ClusterName dimension to the aws.memorydb tag. The NodeId
dimension is already mapped to the nf.node tag.

Considering this mapping, the elasticache aws.node dimension is better
represented by nf.node, which is more in-line with existing common tags.

Add the EngineCPUUtilization metric, which is recommended for use when
there are four or more vCPU.