atlas-persistence: Support common str compression
This adds support for encoding the keys and values in tags as single
character values starting at codepoint 128.

Atlas keys and values are restricted to a subset of the ASCII set so we
can take advantage of this by encoding the most commonly seen strings in
a single unicode character.

If the resource `common-strings.json` is missing we preserve the
previous behavior.