lwc-cloudwatch: reduce frequency of config updates (#378)
When syncing the configs, there will be many updates in
a short window. Before the config manager would forward
the complete set of data sources for each message that is
received. This has a large overhead downstream. Now it
will only push the updated set of data sources when it
receives a done or hearbeat message and there are pending
changes.