aggr: ensure publisher future completes
Make the publisher queue size configurable. Adjust future
to ensure it will always be completed even if the stream
gets restarted. Now it will be completed when returned and
treated as successful if it was enqueued. This avoids
problems with the upstream hanging if the stream is restarted
and some of the returned futures are never completed.