cloudwatch: add FirehoseMetric class (#401)
Represents a metric received from a Kinesis Firehose subscribed
to CloudWatch data. The class uses xxHash for caching the values.
This will be used for a unique identifier in Redis and needs to
have a lower collision possibility due to the number of unique
CloudWatch metrics. It must also be portable so that we can use
other languages if needed.