Atlas Cloud Watch: Move the Tagger function type into a Trait that includes the (#404)
`fixTags()` method. This allows us to move the `fixTags()` from the Client Actor
into the DefaultTagger for re-use when the actor goes away.
Also copy the default valid tag key and value character configs from the sink
to the tagger section in reference.conf as we'll remove the sink section later.