Atlas Cloud Watch: Add the CloudWatchMetricsProcessor, an abstract class (#413)
for writing and reading from a backing store implementation. The processor
is responsible for matching incoming fire hose data to rules and merging
new values with existing values in the cache. It is also responsible
for regularly scraping the cache and publishing the final data.
Included is a simple local hash map based store for unit testing.