atlas cloudwatch: Rewrite CloudWatchPoller. (#436)
This poller now focuses only on polling a small set of metrics once a
day with heavy limiting but for multiple accounts and regions. This
will be used for S3 bucket aggregate metrics and any other series that
are unable to be processed through the Firehose stream due to back-dated
timestamps.