atlas cloudwatch: add AWS Config account supplier. (#444)
The supplier polls an AWS Config database for the list of accounts and resources
active in each region to avoid manual configuration and mappings.
This change re-works the AwsAccountSupplier to return a map of accounts to
regions to namespaces so we can only poll for the data that should be there, saving
a bit of money. Also drops the future from that interface.