atlas cloudwatch: Poll flag per account and region. (#446)
This lets us recover a big more graceffully if a particular account or
region has an issue when polling. Instead of tracking a flag and a key
in redis for the whole job (all accounts and regions) track per account
and region combo so that we can retry only those that fail.
Also set the check for polling state to every 5 minutes instead.