druid: determine type based on aggregator (#456)
There are 2 issues with Druid segment metadata

1.  If a column type is changed, the entry for the column default to a string
   type and an errorMessage is populated. These column entries cannot be
   used. Since we treat all string columns as dimension, a metric that
   changes type gets listed as a dimension instead of a metric and as such
   cannot be queried.
2. The HLLSketchBuilder ingestion aggregator initially reports column type
   of HLLSketchBuilder but as soon as that is persisted into an actual
   segment it becomes HLLSketch. This results in the same symptom as
   (1) since the type appears to have changed.

We encounter problem (1) when someone accidentally adds a new metric
as the wrong type and needs to fix it. e.g. as a spectatorHistogramTimer
when it was supposed to be a spectatorHistogramDistribution. Today we
have to run a hadoop job to restate that column under the correct type.

We encounter problem (2) when ingesting HLLSketches from raw values
using realtime ingestion. This does not appear when using batch ingestion
since the segment and metadata is already finalized once the cluster loads
the segment.

Changes:

1. Filter out columns that report an error. These are typically where a metric
   type has changed between segments. The column type defaults to string,
   but provides a non-null errorMessage. Previously we would assume this
   was a dimension since it is reported as string. This will now omit it from
   visibility.
2. Use aggregators to define metrics rather than columns. Requires using
   lenient aggregator merge if we want to accommodate types changing.