druid: fix OR query using name (#459)
The name is matched before creating the druid filter and
simplified to `true`. If a query has an or clause where
one side is just the name, then it will elide the entire
or clause for the filter.

Now it will simplify the query taking into account the
name match earlier so only the branch of the or for the
particular metric will be present when constructing the
filter.