atlas cloudwatch: Fixes for Media* metrics (#477)
Add support for extracting multiple groups from extractors and combine them
with a hyphen to shorten the MediaConnect source and flow ARNs. Update the
regexes for those configs.
Also remove duplicate SourceARN and FlowARN mappings that were conflicting
with the extractors, causing some metrics to report the fulle string and
others to use the regex.
Bump MediaLive metrics to 4 minute offset.