druid: support data with mixed step sizes (#570)
Update so that some data sources can be more granular than
others and get blended together on the graphs. If a data
source doesn't support the fine granularity, then it will
be retrieved at its supported granularity and mapped to
the requested step for the graph.

This will not work for data sources that are consolidated
historically. That would add quite a bit of complexity to
processing the data, but may be desirable in the future
to help mitigate cost.