druid: adjust the step based on the datasource
For druid the data sources can potentially have different
step sizes. With this change, it will choose the maximum
step size across all potential data sources that could
match an expression on the graph and the computed context
step.

If a higher resolution metric is included on the same
graph as a lower resolution metric, the lower resolution
will be used for both so the data will align for math
and stats.